Share the delimiter functionality between dataset/catalog
