+++ crates/runtime/src/component/mod.rs
@@ -83,6 +83,33 @@ pub fn validate_identifier(identifier: &str) -> Result<(), Error> {
     Ok(())
 }
 
+/// Helper function that finds the position and length of the first delimiter ('://', ':', or '/')
+fn find_first_delimiter(from: &str) -> Option<(usize, usize)> {
+    // Find the earliest occurrence of each delimiter
+    let colon_slash_slash = from.find("://");
+    let colon = from.find(':');
+    let slash = from.find('/');
+
+    // Get the position and length of the first delimiter
+    match (colon_slash_slash, colon, slash) {
+        (Some(css), Some(c), Some(s)) => {
+            let min_pos = css.min(c).min(s);
+            Some(if min_pos == css {
+                (css, 3)
+            } else {
+                (min_pos, 1)
+            })
+        }
+        (Some(css), Some(c), None) => Some(if css < c { (css, 3) } else { (c, 1) }),
+        (Some(css), None, Some(s)) => Some(if css < s { (css, 3) } else { (s, 1) }),
+        (None, Some(c), Some(s)) => Some(if c < s { (c, 1) } else { (s, 1) }),
+        (Some(css), None, None) => Some((css, 3)),
+        (None, Some(c), None) => Some((c, 1)),
+        (None, None, Some(s)) => Some((s, 1)),
+        (None, None, None) => None,
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
